fix: Haskell example end of life?
The `cabal install` command started failing constantly in the Haskell example w/o
any changes to the example's code. Likely, something in the haskel:7
ecosystem reached the end of life.

Bumped up `haskell:7` to `haskell:9` and `cabal:1` to `cabal:2` and
adjusted the Dockerfile a bit to calm down the new cabal version. Now the
example seems to work. However, it caused a quite significant increase
in the image size - the base Haskell image went from 714 MB in haskell:7
to 3.1 GB in haskel:9.